chore(utxo-core): renamed files and moved helper to own file
TICKET: BTC-2047

chore(utxo-core): updated index.ts

TICKET: BTC-2047

chore(utxo-core): added asserts for address in attestation

TICKET: BTC-2047

chore(utxo-core): naming convention

TICKET: BTC-2047
