+++ modules/utxo-core/src/paygo/psbt/index.ts
@@ -1,1 +1,1 @@
-export * from './PayGoUtils';
+export * from './payGoAddressProof';
